setting: project 구조 변경
