+++ boomerang/src/main/java/boomerang/board/controller/BoardController.java
@@ -6,8 +6,8 @@ import boomerang.board.dto.BoardListResponseDto;
 import boomerang.board.dto.BoardRequestDto;
 import boomerang.board.dto.BoardResponseDto;
 import boomerang.board.service.BoardService;
-import boomerang.domain.member.domain.Member;
-import boomerang.domain.member.service.MemberService;
+import boomerang.member.domain.Member;
+import boomerang.member.service.MemberService;
 import boomerang.global.exception.DomainValidationException;
 import boomerang.global.oauth.dto.PrincipalDetails;
 import boomerang.global.response.ErrorResponseDto;
