--- boomerang/src/main/java/boomerang/domain/kakao/domain/KakaoProfile.java
@@ -1,5 +1,0 @@
-package boomerang.domain.kakao.domain;
-
-public record KakaoProfile(Long id, Properties properties) {
-
-}
