--- boomerang/src/main/java/boomerang/domain/kakao/domain/Properties.java
@@ -1,5 +1,0 @@
-package boomerang.domain.kakao.domain;
-
-public record Properties(String nickname) {
-
-}
