--- boomerang/src/main/java/boomerang/domain/kakao/exception/KakaoNotFoundException.java
@@ -1,10 +1,0 @@
-package boomerang.domain.kakao.exception;
-
-import boomerang.global.exception.BusinessException;
-import boomerang.global.response.ErrorCode;
-
-public class KakaoNotFoundException extends BusinessException {
-    public KakaoNotFoundException() {
-        super(ErrorCode.TEMPLATE_NOT_FOUND_ERROR);
-    }
-}
