--- boomerang/src/main/java/boomerang/domain/kakao/exception/KakaoValidException.java
@@ -1,10 +1,0 @@
-package boomerang.domain.kakao.exception;
-
-import boomerang.global.exception.DomainValidationException;
-import boomerang.global.response.ErrorCode;
-
-public class KakaoValidException extends DomainValidationException {
-    public KakaoValidException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
