--- boomerang/src/main/java/boomerang/domain/member/exception/MemberValidException.java
@@ -1,10 +1,0 @@
-package boomerang.domain.member.exception;
-
-import boomerang.global.exception.DomainValidationException;
-import boomerang.global.response.ErrorCode;
-
-public class MemberValidException extends DomainValidationException {
-    public MemberValidException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
