+++ boomerang/src/main/java/boomerang/kakao/domain/KakaoProfile.java
@@ -1,0 +1,5 @@
+package boomerang.kakao.domain;
+
+public record KakaoProfile(Long id, Properties properties) {
+
+}
