+++ boomerang/src/main/java/boomerang/kakao/domain/Properties.java
@@ -1,0 +1,5 @@
+package boomerang.kakao.domain;
+
+public record Properties(String nickname) {
+
+}
