+++ boomerang/src/main/java/boomerang/kakao/dto/KakaoTokenResponseDto.java
@@ -1,0 +1,29 @@
+package boomerang.kakao.dto;
+
+import boomerang.template.domain.TemplateColumn1;
+import boomerang.template.domain.TemplateColumn2;
+import boomerang.template.domain.TemplateDomain;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+
+@Getter
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class KakaoTokenResponseDto {
+
+    @JsonProperty("token_type")
+    public String tokenType;
+    @JsonProperty("access_token")
+    public String accessToken;
+    @JsonProperty("id_token")
+    public String idToken;
+    @JsonProperty("expires_in")
+    public Integer expiresIn;
+    @JsonProperty("refresh_token")
+    public String refreshToken;
+    @JsonProperty("refresh_token_expires_in")
+    public Integer refreshTokenExpiresIn;
+    @JsonProperty("scope")
+    public String scope;
+
+}
