+++ boomerang/src/main/java/boomerang/kakao/exception/KakaoValidException.java
@@ -1,0 +1,10 @@
+package boomerang.kakao.exception;
+
+import boomerang.global.exception.DomainValidationException;
+import boomerang.global.response.ErrorCode;
+
+public class KakaoValidException extends DomainValidationException {
+    public KakaoValidException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
