+++ boomerang/src/main/java/boomerang/member/domain/ReturnDeposit.java
@@ -1,0 +1,49 @@
+package boomerang.member.domain;
+
+import boomerang.member.exception.MemberValidException;
+import boomerang.global.response.ErrorCode;
+import com.fasterxml.jackson.annotation.JsonValue;
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
+import java.util.Objects;
+
+@Embeddable
+public class ReturnDeposit {
+    @Column(name = "return_deposit")
+    private int value;
+
+    public ReturnDeposit() {
+
+    }
+
+    public ReturnDeposit(int value) {
+        // 조건을 통해 valid한 값인지를 체크한다
+        if (false) {
+            throw new MemberValidException(ErrorCode.TEMPLATE_NOT_FOUND_ERROR);
+        }
+        this.value = value;
+    }
+
+    @JsonValue
+    public int getValue() {
+        return value;
+    }
+
+//    @Override
+//    public String toString() {
+//        return value;
+//    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        ReturnDeposit that = (ReturnDeposit) o;
+        return Objects.equals(value, that.value);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(value);
+    }
+}
