+++ boomerang/src/main/java/boomerang/member/dto/MemberCreateResponseDto.java
@@ -1,0 +1,18 @@
+package boomerang.member.dto;
+
+import boomerang.member.domain.Email;
+import lombok.Getter;
+
+@Getter
+public class MemberCreateResponseDto {
+    private Email email;
+
+
+    public MemberCreateResponseDto() {}
+
+    // 생성자
+    public MemberCreateResponseDto(Email email) {
+        this.email = email;
+
+    }
+}
