+++ boomerang/src/main/java/boomerang/member/exception/MemberNotFoundException.java
@@ -1,0 +1,10 @@
+package boomerang.member.exception;
+
+import boomerang.global.exception.BusinessException;
+import boomerang.global.response.ErrorCode;
+
+public class MemberNotFoundException extends BusinessException {
+    public MemberNotFoundException() {
+        super(ErrorCode.TEMPLATE_NOT_FOUND_ERROR);
+    }
+}
