+++ boomerang/src/main/java/boomerang/member/exception/MemberValidException.java
@@ -1,0 +1,10 @@
+package boomerang.member.exception;
+
+import boomerang.global.exception.DomainValidationException;
+import boomerang.global.response.ErrorCode;
+
+public class MemberValidException extends DomainValidationException {
+    public MemberValidException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
